Update version and enhance keypress functionality
Updated version number and modified functions to echo keypress values.

Added the ability for the user to use the pause script to direct key presses to other functions by using command substitution within the script. The script echoes the key press for the user to use it to direct case statements etc.