Refactor pause.sh for improved readability and logic
Changed the $(date +%s) to printf -v VARIABLE '%(%s)T' -1 in order to keep with internal bash commands and improve time calculation to eliminate opening a shell for the seconds, this is more accurate and will eliminate most of the lag time for calculations, also for more backward compatibility with older bash versions. I streamlined the return of options like --prompt and --response in order to work efficiently with command substitution ensuring all options are echoed to their perspective places correctly, that the key pressed is echoed to the variable and the prompt and response are only echoed to the terminal.