Refactor pause.sh for improved user options
Rewrote the countdown function to include both the quiet function as well as the regular countdown. I assured that all options work as intended and double checked compatibility with command substitution for assigning key press value to variable. I  assured the text from the prompt and response go directly to stderr, assuring the prompt/response values can be sent to either log or to terminal/tty.