Refactor time tracking in pause.sh using date command
I was mistaken and used the printf -T function thinking it had more compatibility than the $(date <format>) command but it must have at least version 4.2+ or later to function. Since I have learned that the date command is on almost all Unix-like systems giving the most cross-compatibility. I have changed all the commands to use date for finding if a second has passed  for the countdown timer.