pause-6.0.1.sh
I redid the entire script to streamline and thin down the code to eliminate redundancy. I have changed the handling of the -q, --quiet switch to remove the timer requirement being set. I am relying on the user to make the right choice when to quiet the prompt. I removed the chance of any  piping to eliminate any unwanted behavior. I updated the error handling and also updated the way the script works with unexpected positional parameters and setting an error trap to handle any unexpected exiting and sets all prompt data back to default. All other options remain with the same in functionality.