Enhance pause script with better timer and docs
Updated script documentation and improved timer functionality. Changed the timer function to monotonic comparison by accessing the monotonic system clock instead of the calendar date clock which is not monotonic and can vary going forward and backward and lose time, plus the lag time of calling a child process to run the external program. The value can be extensive over long periods but using the monotonic value from the system clock eliminates lag and time loss. Using the monotonic clock the script can use the comparison of the monotonic values to go for extended periods of time and not lose time because the clock will never be a negative value, always positive.